Add tests and fix reset problem
In order to properly and independently execute the tests, I
had implemented a globals reset procedure. This reset should
restore the globals to a pristine condition, as if freshly
loaded. For some reason this did not work correctly. For the
longest of time I did not note this because the tests just
worked. But then, one test mysteriously failed. If executed
in the REPL it worked flawlessly. I investigated and finally
concluded that the globals reset does not work. I fixed this
and now the tests run successfully.